chore(test): move electron tests to typescript
